refactor reducer to core store
